+++ examples/qvm/expectation.c
@@ -3,6 +3,7 @@
 #include <string.h>
 
 #include "error.h"
+#include "sbcl_librarian.h"
 #include "libquil.h"
 
 void die(char *msg) {
@@ -15,7 +16,7 @@ double do_expectation(quil_program state_prep, quil_program operator) {
   double expectations[1] = {0};
 
   if (qvm_expectation(state_prep, operators, 1, NULL, &expectations) !=
-      LIBQUIL_ERROR_SUCCESS) {
+      LISP_ERR_SUCCESS) {
     LIBQUIL_ERROR("failed to call qvm_expectation");
     exit(1);
   }
@@ -24,21 +25,19 @@ double do_expectation(quil_program state_prep, quil_program operator) {
 }
 
 int main(int argc, char **argv) {
-  init("../../libquil.core");
-
   quil_program i;
   quil_program z;
   quil_program x;
 
-  if (quilc_parse_quil("I 0", &i) != LIBQUIL_ERROR_SUCCESS) {
+  if (quilc_parse_quil("I 0", &i) != LISP_ERR_SUCCESS) {
     LIBQUIL_ERROR("failed to parse quil");
     exit(1);
   }
-  if (quilc_parse_quil("Z 0", &z) != LIBQUIL_ERROR_SUCCESS) {
+  if (quilc_parse_quil("Z 0", &z) != LISP_ERR_SUCCESS) {
     LIBQUIL_ERROR("failed to parse quil");
     exit(1);
   }
-  if (quilc_parse_quil("X 0", &x) != LIBQUIL_ERROR_SUCCESS) {
+  if (quilc_parse_quil("X 0", &x) != LISP_ERR_SUCCESS) {
     LIBQUIL_ERROR("failed to parse quil");
     exit(1);
   }